+++ packages/agents-core/src/agent.ts
@@ -33,6 +33,7 @@ import { RunState } from './runState';
 import { toFunctionToolName } from './utils/tools';
 import { getOutputText } from './utils/messages';
 import { isZodObject } from './utils/typeGuards';
+import { combineAbortSignals } from './utils/abortSignals';
 import { ModelBehaviorError, UserError } from './errors';
 import { RunToolApprovalItem } from './items';
 import logger from './logger';
@@ -170,57 +171,6 @@ type AgentTool<
     handler: AgentToolEventHandler<TAgent>,
   ) => AgentTool<TContext, TAgent, TParameters>;
 };
-
-function combineAbortSignals(...signals: (AbortSignal | undefined)[]): {
-  signal?: AbortSignal;
-  cleanup: () => void;
-} {
-  const activeSignals = signals.filter(Boolean) as AbortSignal[];
-  if (activeSignals.length === 0) {
-    return {
-      cleanup: () => {},
-    };
-  }
-
-  const anyFn = (AbortSignal as any).any;
-  if (typeof anyFn === 'function') {
-    try {
-      return {
-        signal: anyFn(activeSignals),
-        cleanup: () => {},
-      };
-    } catch {
-      // Fall back to manual signal composition for runtimes without AbortSignal.any support.
-    }
-  }
-
-  const controller = new AbortController();
-  const listeners: Array<{ signal: AbortSignal; handler: () => void }> = [];
-  const abortCombined = (reason: unknown) => {
-    if (!controller.signal.aborted) {
-      controller.abort(reason);
-    }
-  };
-
-  for (const signal of activeSignals) {
-    if (signal.aborted) {
-      abortCombined(signal.reason);
-      break;
-    }
-    const handler = () => abortCombined(signal.reason);
-    signal.addEventListener('abort', handler, { once: true });
-    listeners.push({ signal, handler });
-  }
-
-  return {
-    signal: controller.signal,
-    cleanup: () => {
-      for (const listener of listeners) {
-        listener.signal.removeEventListener('abort', listener.handler);
-      }
-    },
-  };
-}
 
 export type ToolUseBehaviorFlags = 'run_llm_again' | 'stop_on_first_tool';
